style(Expeditions)
In a really good spot, need to update the size of the start buttons background & maybe text.
And also need fix the progress bar, we've come down to just text at this point lol.
(Maybe a tiny bit of spacing padding correction, haven't decided yet.
The background 100% needs to change and blend together better as well as the "header".